requirements.txt
torch~=1.8.1+cu111
numpy~=1.19.2
matplotlib~=3.4.1
tensorboardX~=2.2
scipy~=1.6.2
torchvision~=0.9.1+cu111
Pillow~=8.2.0
imageio~=2.9.0
tqdm~=4.60.0